backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "proteosurf"
version = "0.1.0"
description = "Proteosurf — AI-powered structural biology assistant with MCP server"
readme = "../README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]",
    "httpx>=0.27",
    "biopython>=1.84",
    "numpy>=1.26",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "websockets>=13.0",
]

[project.optional-dependencies]
docking = ["rdkit"]
voice = ["elevenlabs>=2.0"]
databricks = ["mlflow>=2.0", "databricks-sdk"]
nemotron = []
nia = ["nia-ai-py"]
claude-agent = ["claude-agent-sdk>=0.1.30"]
all = [
    "rdkit",
    "elevenlabs>=2.0",
    "mlflow>=2.0",
    "databricks-sdk",
    "nia-ai-py",
    "claude-agent-sdk>=0.1.30",
]

[project.scripts]
proteosurf-mcp = "proteosurf.mcp_server:main"
proteosurf-web = "proteosurf.run_web:main"

[tool.hatch.build.targets.wheel]
packages = ["proteosurf"]
